Undo refactor (as it's now only used in one place)
