+++ datahub/search/views.py
@@ -1,7 +1,6 @@
 """Search views."""
 from collections import namedtuple
 from enum import auto, Enum
-from functools import reduce
 from itertools import islice
 
 from django.conf import settings
@@ -290,46 +289,6 @@ class SearchAPIView(APIView):
     def get_serializer(self):
         """Return query serializer for use with OpenAPI documentation."""
         return self.serializer_class()
-
-    def deep_get(self, dictionary, keys, default=None):
-        """
-        Perform a deep search on a dictionary to find the item at the location provided in the keys
-        """
-        return reduce(
-            lambda d, key: d.get(key, default) if isinstance(d, dict) else default,
-            keys.split('|'),
-            dictionary,
-        )
-
-    def add_must_and_must_not_to_filters(self, base_query, must, must_not):
-        """
-        Helper method to add must and must_not filters to the raw_query.
-        """
-        raw_query = base_query.to_dict()
-        filters = self.deep_get(raw_query, 'query|bool|filter')
-        if not filters:
-            return raw_query
-
-        filter_index = None
-        for index, filter in enumerate(filters):
-            if filter.get('bool') or filter.get('bool') == {}:
-                filter_index = index
-                break
-
-        if filter_index is None:
-            return raw_query
-
-        if len(must_not) > 0:
-            filters[filter_index]['bool']['must_not'] = must_not
-        if len(must) > 0:
-            if 'must' not in filters[filter_index]['bool']:
-                filters[filter_index]['bool']['must'] = []
-
-            filters[filter_index]['bool']['must'].append(must)
-
-        raw_query['query']['bool']['filter'] = filters
-
-        return raw_query
 
 
 class SearchExportAPIView(SearchAPIView):
